requirements.txt
python-telegram-bot==20.7
google-genai
pdfplumber
fastapi
uvicorn
pydantic
python-multipart
aiofiles
python-dotenv
easyocr
pillow
opencv-python-headless
numpy
